[charts] Performance tests: set license on setup. Update vitest minor version.
