+++ test/performance-charts/package.json
@@ -16,8 +16,8 @@
     "@testing-library/user-event": "^14.6.1",
     "@vitejs/plugin-react": "^4.3.4",
     "@vitejs/plugin-react-swc": "^3.7.2",
-    "@vitest/browser": "3.0.4",
-    "@vitest/ui": "3.0.4",
+    "@vitest/browser": "3.0.5",
+    "@vitest/ui": "3.0.5",
     "jsdom": "^25.0.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
