+++ test/performance-charts/tests/BarChartPro.bench.tsx
@@ -3,7 +3,6 @@ import * as React from 'react';
 import { render, cleanup } from '@testing-library/react';
 import { bench, describe } from 'vitest';
 import { BarChartPro } from '@mui/x-charts-pro/BarChartPro';
-import { LicenseInfo, generateLicense } from '@mui/x-license';
 import { options } from '../utils/options';
 
 describe('BarChartPro', () => {
@@ -19,16 +18,6 @@ describe('BarChartPro', () => {
   bench(
     'BarChartPro with big data amount',
     async () => {
-      const licenseKey = generateLicense({
-        expiryDate: new Date(3001, 0, 0, 0, 0, 0, 0),
-        orderNumber: 'MUI-123',
-        planScope: 'pro',
-        licenseModel: 'subscription',
-        planVersion: 'Q3-2024',
-      });
-
-      LicenseInfo.setLicenseKey(licenseKey);
-
       const { findByText } = render(
         <BarChartPro
           xAxis={[{ id: 'x', scaleType: 'band', data: xData, zoom: { filterMode: 'discard' } }]}
