+++ test/performance-charts/vitest.config.ts
@@ -5,6 +5,7 @@ import react from '@vitejs/plugin-react';
 export default defineConfig({
   plugins: [codspeedPlugin(), react()],
   test: {
+    setupFiles: ['./setup.ts'],
     environment: 'jsdom',
     // browser: {
     //   enabled: true,
